Add files via upload
Added a file, which is from Jerome's commit yesterday which fixed the calculation of duration. I converted the True/False setting for the receptive_field_mapping as an input to the function, and then added gabor_rf with true as well as false. After reducing some trials for the bars of light, I am getting 10:41 as run time for 1 each (true then false for shuffling) of Gabor_rf and N_REPEATS=5.